Make TopAppBar Reusable
