+++ src/data/NavBarTypes.ts
@@ -1,0 +1,2 @@
+export type NavBarItemRecord = { title: string; link: string };
+export type NestedRecord = { [k: string]: Array<NavBarItemRecord> | NestedRecord };
